Add files via upload
This is a Jupyter notebook file named DataScienceEcosystem.ipynb created as a course requirement for final peer review assingment. Course 2 of 10 IBM Data Science Professional Certificate course. 
In the file, markdown and code cells have been created and several functionalities of the Jupyter Notebook along with Python have been practiced as a part of final assignment.